requirements.txt
asgiref==3.3.4
attrs==21.4.0
boto3==1.18.30
botocore==1.21.30
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
oauthlib==3.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.11.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
whitenoise==5.3.0
